부등호 문제 디렉토리 이동
